DWARF parsing library
This is complete enough to recursively scan the entire .debug_info
section of the Linux kernel.  This requires full parsing of the abbrev
table, structural parsing of DIEs in the info table, and support for
following sibling attributes.  It does not yet support other debug
sections or decoding attribute value other than references and there
are many blatant inefficiencies.