Fix local pkg-configs
Commit f8eac5a changed the install and the pkg-config to use proper
library and include paths, but added these to the *local* pkg-config,
rather than introducing them at the point when the pkg-config was
installed. As a result, building against the uninstalled .pc file
(such as what mtrace-tools does when cloned next to libelfin) no
longer worked because it referred to the wrong paths.

Fix this by reverting the changes to the local .pc file and
introducing the lib and include paths only when installing the .pc
file globally.