elf: fix Shdr.link for big-endian byte order
The section number type is declared as a Half (uint16_t) because
that's the right size for Sym, but the link field section number in
Shdr is actually a Word (uint32_t). This works out find on
little-endian because the padding inserted by the compiler is in the
right place, but on big-endian the padding is on the wrong side of
this field, so the link field is always 0 in practice.

Fix this by using the real "shn" type if the byte order is "native",
and using Word if the byte order is big or little endian,

Updates #19.