ignore dwarf extensions of gcc in the line table
gcc 9 uses DWARF 4 by default with vendor extensions in the line table which
results in a thrown exception. An application using libelfin cannot recover
from an exception and still get a reasonable line table. Now, when a vendor
extension is encountered, the type is set to invalid to get around the issue.