Add DESTDIR support
For packaging it is useful to be able to install to a custom temporary
installation location. The existing PREFIX variable is not enough for
this purpose because it will lead to incorrect pkg-config files (e.g.
setting PREFIX="$pkgdir/usr" leads to includedir=$pkgdir/usr/include
while it should be includedir=/usr/include instead). DESTDIR is the
standard name for this variable, defined by the GNU coding standards.